+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Raising Exceptions with `movements.py` and `cli.py`
 - Cognitive Complexity in `utils.py`
 - `Dockerfile` adjusted to the project
+- Tests now can be executed
 
 ### Changed